+++ src/JsonApi/Exception/RelationshipTypeNotAppropriate.php
@@ -14,17 +14,24 @@ class RelationshipTypeNotAppropriate extends \Exception
     private $relationshipType;
 
     /**
+     * @var string
+     */
+    private $expectedRelationshipType;
+
+    /**
      * @param string $relationshipName
      * @param string $relationshipType
+     * @param string $expectedRelationshipType
      */
-    public function __construct($relationshipName, $relationshipType)
+    public function __construct($relationshipName, $relationshipType, $expectedRelationshipType)
     {
         parent::__construct(
-            "The provided relationship '$relationshipName' is of type of $relationshipType, " .
-            "but it is inappropriate."
+            "The provided relationship '$relationshipName' is of type of $relationshipType, but " .
+            ($expectedRelationshipType ? "$expectedRelationshipType is" : "it is not the one which is") . " expected."
         );
         $this->relationshipName = $relationshipName;
         $this->relationshipType = $relationshipType;
+        $this->expectedRelationshipType = $expectedRelationshipType;
     }
 
     /**
@@ -41,5 +48,13 @@ class RelationshipTypeNotAppropriate extends \Exception
     public function getRelationshipType()
     {
         return $this->relationshipType;
+    }
+
+    /**
+     * @return string
+     */
+    public function getExpectedRelationshipType()
+    {
+        return $this->expectedRelationshipType;
     }
 }